+++ src/lab14.py
@@ -1,838 +1,856 @@
 """
 This file compiles the code in Web Browser Engineering,
-up to and including Chapter 14 (Animations and Compositing),
+up to and including Chapter 13 (Saving Partial Layouts),
 without exercises.
 """
 
-import ctypes
-import dukpy
-import io
-import math
-import sdl2
-import sdl2.ext as sdl2ext
-import skia
 import socket
 import ssl
-import threading
+import tkinter
+import tkinter.font
+import dukpy
 import time
-import urllib.parse
-from lab4 import print_tree
-from lab4 import Element
-from lab4 import Text
-from lab4 import HTMLParser
-from lab6 import cascade_priority
-from lab6 import layout_mode
-from lab6 import resolve_url
-from lab6 import tree_to_list
-from lab6 import INHERITED_PROPERTIES
-from lab6 import compute_style
-from lab6 import TagSelector, DescendantSelector
-from lab9 import EVENT_DISPATCH_CODE
-from lab10 import COOKIE_JAR, request, url_origin
-from lab11 import get_font, linespace, parse_blend_mode, parse_color
-from lab12 import raster
+from lab10 import request
 
 class Timer:
     def __init__(self):
+        self.phase = None
         self.time = None
 
-    def start(self):
+    def start(self, name):
+        if self.phase: self.stop()
+        self.phase = name
         self.time = time.time()
 
     def stop(self):
-        return time.time() - self.time
-        self.time = None
+        dt = time.time() - self.time
+        print("[{:>10.6f}] {}".format(dt, self.phase))
+        self.phase = None
 
-FONTS = {}
-
-def async_request(url, top_level_url, results):
-    headers = None
-    body = None
-    def runner():
-        headers, body = request(url, top_level_url)
-        results[url] = {'headers': headers, 'body': body}
-    thread = threading.Thread(target=runner)
-    thread.start()
-    return thread
-
-def center_point(rect):
-    return (rect.left() + (rect.right() - rect.left()) / 2,
-        rect.top() + (rect.bottom() - rect.top()) / 2)
-
-class DisplayItem:
-    def __init__(self, rect, needs_compositing=False, cmds=None,
-        is_noop=False):
-        self.rect = rect if rect else skia.Rect.MakeEmpty()
-        self.composited=needs_compositing
-        self.cmds = cmds
-        self.noop = is_noop
-
-    def bounds(self):
-        return self.rect
-
-    def needs_compositing(self):
-        if self.composited:
-            return True
-        return False
-
-    def get_cmds(self):
-        return self.cmds
-
-    def is_noop(self):
-        return self.noop
-
-    def execute(self, canvas):
-        if self.cmds:
-            def op():
-                for cmd in self.get_cmds():
-                    cmd.execute(canvas, True, False)
-            self.draw(canvas, op, True)
-        else:
-            self.draw(canvas)
-
-    def draw(self, canvas, op, is_raster):
-        pass
-
-    def transform(self, rect):
-        return rect
-
-    def repr_recursive(self, indent=0, include_noop=False):
-        inner = ""
-        if not include_noop and self.is_noop():
-            if self.cmds:
-                for cmd in self.cmds:
-                   inner += cmd.repr_recursive(indent, include_noop)
-                return inner
-        else:
-            if self.cmds:
-                for cmd in self.cmds:
-                    inner += cmd.repr_recursive(indent + 2, include_noop)
-            return ("{indentation}{repr}: bounds={bounds}, " +
-                "needs_compositing={needs_compositing}{noop}\n{inner} ").format(
-                indentation=" " * indent,
-                repr=self.__repr__(),
-                bounds=self.bounds(),
-                needs_compositing=self.needs_compositing(),
-                inner=inner,
-                noop=(" <no-op>" if self.is_noop() else ""))
-
-class Transform(DisplayItem):
-    def __init__(self, translation, rotation_degrees, rect, cmds,
-        should_transform):
-        self.rotation_degrees = rotation_degrees
-        self.translation = translation
-        (self.center_x, self.center_y) = center_point(rect)
-        assert translation == None or rotation_degrees == None
-        self.should_transform = should_transform
-        my_bounds = self.compute_bounds(rect, cmds, should_transform)
-        super().__init__(my_bounds, should_transform, cmds, not should_transform)
-
-    def draw(self, canvas, op, is_raster):
-        if self.is_noop():
-            op()
-        elif self.translation:
-            (x, y) = self.translation
-            canvas.save()
-            canvas.translate(x, y)
-            op()
-            canvas.restore()
-        else:
-            rotation_x = self.center_x
-            rotation_y = self.center_y
-            if not is_raster:
-                rotation_x -= self.bounds().left()
-                rotation_y -= self.bounds().top()
-            canvas.save()
-            canvas.rotate(
-                degrees=self.rotation_degrees, px=rotation_x, py=rotation_y)
-            op()
-            canvas.restore()
-
-    def transform(self, rect):
-        return self.transform_internal(rect, not self.is_noop())
-
-    def transform_internal(self, rect, should_transform):
-        if not should_transform:
-            return rect
-        matrix = skia.Matrix()
-        if self.translation:
-            (x, y) = self.translation
-            matrix.setTranslate(x, y)
-        else:
-            matrix.setRotate(
-                self.rotation_degrees, self.center_x, self.center_y)
-        return matrix.mapRect(rect)
-
-    def compute_bounds(self, rect, cmds, should_transform):
-        for cmd in cmds:
-            rect.join(cmd.bounds())
-        return self.transform_internal(rect, should_transform)
-
-    def __repr__(self):
-        if self.is_noop():
-            return "Transform(<no-op>)"
-        elif self.translation:
-            return "Transform(translate({}, {}))".format(self.translation)
-        else:
-            return "Transform(rotate({}))".format(self.rotation_degrees)
-
-class DrawRRect(DisplayItem):
-    def __init__(self, rect, radius, color):
-        self.rrect = skia.RRect.MakeRectXY(rect, radius, radius)
-        self.color = color
-        super().__init__(rect)
-
-    def draw(self, canvas):
-        sk_color = parse_color(self.color)
-        canvas.drawRRect(self.rrect,
-            paint=skia.Paint(Color=sk_color))
-
-    def print(self, indent=0):
-        return " " * indent + self.__repr__()
-
-    def __repr__(self):
-        return "DrawRRect(rect={}, color={})".format(
-            str(self.rrect), self.color)
-
-class DrawText(DisplayItem):
-    def __init__(self, x1, y1, text, font, color):
-        self.left = x1
-        self.top = y1
-        self.right = x1 + font.measureText(text)
-        self.bottom = y1 - font.getMetrics().fAscent + font.getMetrics().fDescent
-        self.rect = \
-        self.font = font
+class Text:
+    def __init__(self, text):
         self.text = text
-        self.color = color
-        super().__init__(skia.Rect.MakeLTRB(x1, y1, self.right, self.bottom))
-
-    def draw(self, canvas, **args):
-        draw_text(canvas, self.left, self.top,
-            self.text, self.font, self.color)
 
     def __repr__(self):
-        return "DrawText(text={})".format(self.text)
+        return "\"" + self.text.replace("\n", "\\n") + "\""
 
-class DrawRect(DisplayItem):
-    def __init__(self, x1, y1, x2, y2, color):
-        self.top = y1
-        self.left = x1
-        self.bottom = y2
-        self.right = x2
-        self.color = color
-        super().__init__(skia.Rect.MakeLTRB(x1, y1, x2, y2))
+SELF_CLOSING_TAGS = [
+    "area", "base", "br", "col", "embed", "hr", "img", "input",
+    "link", "meta", "param", "source", "track", "wbr",
+]
 
-    def draw(self, canvas, *args):
-        draw_rect(canvas,
-            self.left, self.top,
-            self.right, self.bottom,
-            fill=self.color, width=0)
-
-    def __repr__(self):
-        return "DrawRect(top={} left={} bottom={} right={} color={})".format(
-            self.left, self.top, self.right, self.bottom, self.color)
-
-class ClipRRect(DisplayItem):
-    def __init__(self, rect, radius, cmds, should_clip=True):
-        self.rrect = skia.RRect.MakeRectXY(rect, radius, radius)
-        super().__init__(
-            ClipRRect.compute_bounds(rect, cmds), False, cmds, not should_clip)
-
-    def draw(self, canvas, op, is_raster):
-        if not self.is_noop():
-            canvas.save()
-            canvas.clipRRect(self.rrect)
-        op()
-        if not self.is_noop():
-            canvas.restore()
-
-    def compute_bounds(rect, cmds):
-        for cmd in cmds:
-            rect.join(cmd.bounds())
-        return rect
+class Tag:
+    def __init__(self, text):
+        parts = text.split()
+        self.tag = parts[0].lower()
+        self.attributes = {}
+        for attrpair in parts[1:]:
+            if "=" in attrpair:
+                key, value = attrpair.split("=", 1)
+                if len(value) > 2 and value[0] in ["'", "\""]:
+                    value = value[1:-1]
+                self.attributes[key.lower()] = value
+            else:
+                self.attributes[attrpair.lower()] = ""
 
     def __repr__(self):
-        if self.is_noop():
-            return "ClipRRect(<no-op>)"
+        return "<" + self.tag + ">"
+
+def lex(body):
+    out = []
+    text = ""
+    in_tag = False
+    for c in body:
+        if c == "<":
+            in_tag = True
+            if text: out.append(Text(text))
+            text = ""
+        elif c == ">":
+            in_tag = False
+            out.append(Tag(text))
+            text = ""
         else:
-            return "ClipRRect({})".format(str(self.rrect))
+            text += c
+    if not in_tag and text:
+        out.append(Text(text))
+    return out
 
-class DrawLine(DisplayItem):
-    def __init__(self, x1, y1, x2, y2):
-        self.x1 = x1
-        self.y1 = y1
-        self.x2 = x2
-        self.y2 = y2
-        super().__init__(skia.Rect.MakeLTRB(x1, y1, x2, y2))
+class ElementNode:
+    def __init__(self, tag, attributes):
+        self.tag = tag
+        self.attributes = attributes
+        self.children = []
 
-    def execute(self, canvas, *args):
-        draw_line(canvas, self.x1, self.y1, self.x2, self.y2)
-
-class SaveLayer(DisplayItem):
-    def __init__(self, sk_paint, cmds,
-            should_save=True, should_paint_cmds=True, needs_animation=False):
-        self.should_paint_cmds = should_paint_cmds
-        self.sk_paint = sk_paint
-        rect = skia.Rect.MakeEmpty()
-        for cmd in cmds:
-            rect.join(cmd.rect)
-        super().__init__(rect, False, cmds, not should_save)
-
-    def draw(self, canvas, op, is_raster):
-        if not self.is_noop():
-            canvas.saveLayer(paint=self.sk_paint)
-        if self.should_paint_cmds:
-            op()
-        if not self.is_noop():
-            canvas.restore()
+        self.style = {}
+        for pair in self.attributes.get("style", "").split(";"):
+            if ":" not in pair: continue
+            prop, val = pair.split(":")
+            self.style[prop.strip().lower()] = val.strip()
 
     def __repr__(self):
-        if self.is_noop():
-            return "SaveLayer(<no-op>)"
+        return "<" + self.tag + ">"
+
+class TextNode:
+    def __init__(self, text):
+        self.text = text
+        self.tag = None
+        self.children = []
+
+    def __repr__(self):
+        return self.text.replace("\n", "\\n")
+        
+def parse(tokens):
+    currently_open = []
+    for tok in tokens:
+        implicit_tags(tok, currently_open)
+        if isinstance(tok, Text):
+            node = TextNode(tok.text)
+            if not currently_open: continue
+            node.parent = currently_open[-1]
+            currently_open[-1].children.append(node)
+        elif tok.tag.startswith("/"):
+            node = currently_open.pop()
+            if not currently_open: return node
+            currently_open[-1].children.append(node)
+        elif tok.tag in SELF_CLOSING_TAGS:
+            node = ElementNode(tok.tag, tok.attributes)
+            node.parent = currently_open[-1]
+            currently_open[-1].children.append(node)
+        elif tok.tag.startswith("!"):
+            continue
         else:
-            return "SaveLayer(alpha={})".format(self.sk_paint.getAlphaf())
+            node = ElementNode(tok.tag, tok.attributes)
+            node.parent = currently_open[-1]
+            currently_open.append(node)
+    while currently_open:
+        node = currently_open.pop()
+        if not currently_open: return node
+        currently_open[-1].children.append(node)
 
-def draw_line(canvas, x1, y1, x2, y2):
-    path = skia.Path().moveTo(x1, y1).lineTo(x2, y2)
-    paint = skia.Paint(Color=skia.ColorBLACK)
-    paint.setStyle(skia.Paint.kStroke_Style)
-    paint.setStrokeWidth(1);
-    canvas.drawPath(path, paint)
-
-def draw_text(canvas, x, y, text, font, color=None):
-    sk_color = parse_color(color)
-    paint = skia.Paint(AntiAlias=True, Color=sk_color)
-    canvas.drawString(
-        text, float(x), y - font.getMetrics().fAscent,
-        font, paint)
-
-def draw_rect(canvas, l, t, r, b, fill=None, width=1):
-    paint = skia.Paint()
-    if fill:
-        paint.setStrokeWidth(width);
-        paint.setColor(parse_color(fill))
-    else:
-        paint.setStyle(skia.Paint.kStroke_Style)
-        paint.setStrokeWidth(1);
-        paint.setColor(skia.ColorBLACK)
-    rect = skia.Rect.MakeLTRB(l, t, r, b)
-    canvas.drawRect(rect, paint)
-
-def parse_rotation_transform(transform_str):
-    left_paren = transform_str.find('(')
-    right_paren = transform_str.find('deg)')
-    return float(transform_str[left_paren + 1:right_paren])
-
-def parse_translate_transform(transform_str):
-    left_paren = transform_str.find('(')
-    right_paren = transform_str.find(')')
-    (x_px, y_px) = \
-        transform_str[left_paren + 1:right_paren].split(",")
-    return (float(x_px[:-2]), float(y_px[:-2]))
-
-def parse_transform(transform_str):
-    if transform_str.find('translate') >= 0:
-        return (parse_translate_transform(transform_str), None)
-    elif transform_str.find('rotate') >= 0:
-        return (None, parse_rotation_transform(transform_str))
-    else:
-        return (None, None)
+HEAD_TAGS = [
+    "base", "basefont", "bgsound", "noscript",
+    "link", "meta", "title", "style", "script",
+]
+            
+def implicit_tags(tok, currently_open):
+    tag = tok.tag if isinstance(tok, Tag) else None
+    while True:
+        open_tags = [node.tag for node in currently_open]
+        if open_tags == [] and tag != "html":
+            node = ElementNode("html", {})
+            node.parent = None
+            currently_open.append(node)
+        elif open_tags == ["html"] and tag not in ["head", "body", "/html"]:
+            if tag in HEAD_TAGS:
+                implicit = "head"
+            else:
+                implicit = "body"
+            node = ElementNode(implicit, {})
+            node.parent = currently_open[-1]
+            currently_open.append(node)
+        elif open_tags == ["html", "head"] and tag not in ["/head"] + HEAD_TAGS:
+            node = currently_open.pop()
+            currently_open[-1].children.append(node)
+        else:
+            break
 
 class CSSParser:
     def __init__(self, s):
         self.s = s
-        self.i = 0
 
-    def whitespace(self):
-        while self.i < len(self.s) and self.s[self.i].isspace():
-            self.i += 1
+    def whitespace(self, i):
+        while i < len(self.s) and self.s[i].isspace():
+            i += 1
+        return None, i
 
-    def literal(self, literal):
-        assert self.i < len(self.s) and self.s[self.i] == literal
-        self.i += 1
+    def literal(self, i, literal):
+        l = len(literal)
+        assert self.s[i:i+l] == literal
+        return None, i + l
 
-    def word(self):
-        start = self.i
-        in_quote = False
-        while self.i < len(self.s):
-            cur = self.s[self.i]
-            if cur == "'":
-                in_quote = not in_quote
-            if cur.isalnum() or cur in ",/#-.%()\"'" \
-                or (in_quote and cur == ':'):
-                self.i += 1
-            else:
-                break
-        assert self.i > start
-        return self.s[start:self.i]
+    def word(self, i):
+        j = i
+        while j < len(self.s) and self.s[j].isalnum() or self.s[j] in "-.":
+            j += 1
+        assert j > i
+        return self.s[i:j], j
 
-    def pair(self):
-        prop = self.word()
-        self.whitespace()
-        self.literal(":")
-        self.whitespace()
-        val = self.word()
-        return prop.lower(), val
+    def pair(self, i):
+        prop, i = self.word(i)
+        _, i = self.whitespace(i)
+        _, i = self.literal(i, ":")
+        _, i = self.whitespace(i)
+        val, i = self.word(i)
+        return (prop.lower(), val), i
 
-    def ignore_until(self, chars):
-        while self.i < len(self.s):
-            if self.s[self.i] in chars:
-                return self.s[self.i]
-            else:
-                self.i += 1
+    def ignore_until(self, i, chars):
+        while i < len(self.s) and self.s[i] not in chars:
+            i += 1
+        return None, i
 
-    def body(self):
+    def body(self, i):
         pairs = {}
-        while self.i < len(self.s) and self.s[self.i] != "}":
+        _, i = self.literal(i, "{")
+        _, i = self.whitespace(i)
+        while i < len(self.s) and self.s[i] != "}":
             try:
-                prop, val = self.pair()
-                pairs[prop.lower()] = val
-                self.whitespace()
-                self.literal(";")
-                self.whitespace()
+                (prop, val), i = self.pair(i)
+                pairs[prop] = val
+                _, i = self.whitespace(i)
+                _, i = self.literal(i, ";")
             except AssertionError:
-                why = self.ignore_until([";", "}"])
-                if why == ";":
-                    self.literal(";")
-                    self.whitespace()
-                else:
-                    break
-        return pairs
+                _, i = self.ignore_until(i, [";", "}"])
+                if i < len(self.s) and self.s[i] == ";":
+                    _, i = self.literal(i, ";")
+            _, i = self.whitespace(i)
+        _, i = self.literal(i, "}")
+        return pairs, i
 
-    def selector(self):
-        out = TagSelector(self.word().lower())
-        self.whitespace()
-        while self.i < len(self.s) and self.s[self.i] != "{":
-            tag = self.word()
-            descendant = TagSelector(tag.lower())
-            out = DescendantSelector(out, descendant)
-            self.whitespace()
-        return out
+    def selector(self, i):
+        if self.s[i] == "#":
+            _, i = self.literal(i, "#")
+            name, i = self.word(i)
+            return IdSelector(name), i
+        elif self.s[i] == ".":
+            _, i = self.literal(i, ".")
+            name, i = self.word(i)
+            return ClassSelector(name), i
+        else:
+            name, i = self.word(i)
+            return TagSelector(name.lower()), i
+
+    def rule(self, i):
+        selector, i = self.selector(i)
+        _, i = self.whitespace(i)
+        body, i = self.body(i)
+        return (selector, body), i
+
+    def file(self, i):
+        rules = []
+        _, i = self.whitespace(i)
+        while i < len(self.s):
+            try:
+                rule, i = self.rule(i)
+                rules.append(rule)
+            except AssertionError:
+                _, i = self.ignore_until(i, "}")
+                _, i = self.literal(i, "}")
+            _, i = self.whitespace(i)
+        return rules, i
 
     def parse(self):
-        rules = []
-        while self.i < len(self.s):
-            try:
-                self.whitespace()
-                selector = self.selector()
-                self.literal("{")
-                self.whitespace()
-                body = self.body()
-                self.literal("}")
-                rules.append((selector, body))
-            except AssertionError:
-                why = self.ignore_until(["}"])
-                if why == "}":
-                    self.literal("}")
-                    self.whitespace()
-                else:
-                    break
+        rules, _ = self.file(0)
         return rules
+    
+class TagSelector:
+    def __init__(self, tag):
+        self.tag = tag
 
-class BlockLayout:
-    def __init__(self, node, parent, previous):
+    def matches(self, node):
+        return self.tag == node.tag
+
+    def priority(self):
+        return 1
+
+class ClassSelector:
+    def __init__(self, cls):
+        self.cls = cls
+
+    def matches(self, node):
+        return self.cls in node.attributes.get("class", "").split()
+
+    def priority(self):
+        return 16
+
+class IdSelector:
+    def __init__(self, id):
+        self.id = id
+
+    def matches(self, node):
+        return self.id == node.attributes.get("id", "")
+
+    def priority(self):
+        return 256
+
+INHERITED_PROPERTIES = {
+    "font-style": "normal",
+    "font-weight": "normal",
+    "font-size": "16px",
+    "color": "black",
+}
+
+def style(node, parent, rules):
+    if isinstance(node, TextNode):
+        node.style = parent.style
+    else:
+        for selector, pairs in rules:
+            if selector.matches(node):
+                for property in pairs:
+                    if property not in node.style:
+                        node.style[property] = pairs[property]
+        for property, default in INHERITED_PROPERTIES.items():
+            if property not in node.style:
+                if parent:
+                    node.style[property] = parent.style[property]
+                else:
+                    node.style[property] = default
+    for child in node.children:
+        style(child, node, rules)
+
+WIDTH, HEIGHT = 800, 600
+HSTEP, VSTEP = 13, 18
+
+SCROLL_STEP = 100
+
+class LineLayout:
+    def __init__(self, node, parent):
         self.node = node
         self.parent = parent
-        self.previous = previous
+        self.cx = 0
+        self.laid_out = False
         self.children = []
-        self.x = None
-        self.y = None
-        self.width = None
-        self.height = None
 
-    def layout(self):
-        previous = None
-        for child in self.node.children:
-            if layout_mode(child) == "inline":
-                next = InlineLayout(child, self, previous)
-            else:
-                next = BlockLayout(child, self, previous)
-            self.children.append(next)
-            previous = next
+    def append(self, child):
+        self.children.append(child)
+        child.parent = self
+        self.cx += child.w + child.font.measure(" ")
 
-        self.width = style_length(
-            self.node, "width", self.parent.width)
-        self.x = self.parent.x
+    def size(self):
+        self.w = self.parent.w
+        self.compute_height()
 
-        if self.previous:
-            self.y = self.previous.y + self.previous.height
+    def compute_height(self):
+        if not self.children:
+            self.h = 0
+            return
+        self.metrics = [child.font.metrics() for child in self.children]
+        self.max_ascent = max([metric["ascent"] for metric in self.metrics])
+        self.max_descent = max([metric["descent"] for metric in self.metrics])
+        self.h = 1.25 * (self.max_descent + self.max_ascent)
+
+        cx = 0
+        self.cxs = []
+        for child in self.children:
+            self.cxs.append(cx)
+            cx += child.w + child.font.measure(" ")
+
+    def position(self):
+        baseline = self.y + 1.25 * self.max_ascent
+        for cx, child, metrics in \
+          zip(self.cxs, self.children, self.metrics):
+            child.x = self.x + cx
+            child.y = baseline - metrics["ascent"]
+
+    def draw(self, to):
+        for child in self.children:
+            child.draw(to)
+
+class TextLayout:
+    def __init__(self, node, word):
+        self.node = node
+        self.children = []
+        self.word = word
+
+    def size(self):
+        weight = self.node.style["font-weight"]
+        style = self.node.style["font-style"]
+        if style == "normal": style = "roman"
+        size = int(px(self.node.style["font-size"]) * .75)
+        self.font = tkinter.font.Font(size=size, weight=weight, slant=style)
+        
+        self.w = self.font.measure(self.word)
+        self.compute_height()
+
+    def compute_height(self):
+        self.h = self.font.metrics('linespace')
+
+    def position(self):
+        pass
+
+    def draw(self, to):
+        color = self.node.style["color"]
+        to.append(DrawText(self.x, self.y, self.word, self.font, color))
+
+class InputLayout:
+    def __init__(self, node):
+        self.node = node
+        self.children = []
+
+    def size(self):
+        weight = self.node.style["font-weight"]
+        style = self.node.style["font-style"]
+        if style == "normal": style = "roman"
+        size = int(px(self.node.style["font-size"]) * .75)
+        self.font = tkinter.font.Font(size=size, weight=weight, slant=style)
+        self.w = 200
+        self.compute_height()
+
+    def compute_height(self):
+        self.h = 20
+
+    def position(self):
+        pass
+
+    def draw(self, to):
+        x1, x2 = self.x, self.x + self.w
+        y1, y2 = self.y, self.y + self.h
+        bgcolor = "light gray" if self.node.tag == "input" else "yellow"
+        to.append(DrawRect(x1, y1, x2, y2, bgcolor))
+
+        if self.node.tag == "input":
+            text = self.node.attributes.get("value", "")
         else:
-            self.y = self.parent.y
-
-        for child in self.children:
-            child.layout()
-
-        self.height = style_length(
-            self.node, "height",
-            sum([child.height for child in self.children]))
-
-    def paint(self, display_list):
-        cmds = []
-
-        rect = skia.Rect.MakeLTRB(
-            self.x, self.y,
-            self.x + self.width, self.y + self.height)
-        bgcolor = self.node.style.get("background-color",
-                                 "transparent")
-        if bgcolor != "transparent":
-            radius = float(
-                self.node.style.get("border-radius", "0px")[:-2])
-            cmds.append(DrawRRect(rect, radius, bgcolor))
-
-        for child in self.children:
-            child.paint(cmds)
-
-        cmds = paint_visual_effects(self.node, cmds, rect)
-        display_list.extend(cmds)
-
-    def __repr__(self):
-        return "BlockLayout(x={}, y={}, width={}, height={})".format(
-            self.x, self.x, self.width, self.height)
+            text = self.node.children[0].text
+        color = self.node.style["color"]
+        to.append(DrawText(self.x, self.y, text, self.font, color))
 
 class InlineLayout:
-    def __init__(self, node, parent, previous):
+    def __init__(self, node, parent):
         self.node = node
         self.parent = parent
-        self.previous = previous
-        self.children = []
-        self.x = None
-        self.y = None
-        self.width = None
-        self.height = None
-        self.display_list = None
 
-    def layout(self):
-        self.width = style_length(
-            self.node, "width", self.parent.width)
+    def size(self):
+        self.children = [LineLayout(self.node, self)]
+        self.mt = self.bt = self.pt = 0
+        self.mr = self.br = self.pr = 0
+        self.mb = self.bb = self.pb = 0
+        self.ml = self.bl = self.pl = 0
 
-        self.x = self.parent.x
+        self.w = self.parent.w - self.parent.pl - self.parent.pr \
+            - self.parent.bl - self.parent.br
 
-        if self.previous:
-            self.y = self.previous.y + self.previous.height
-        else:
-            self.y = self.parent.y
-
-        self.new_line()
         self.recurse(self.node)
-        
-        for line in self.children:
-            line.layout()
+        self.flush()
+        self.children.pop()
+        self.compute_height()
 
-        self.height = style_length(
-            self.node, "height",
-            sum([line.height for line in self.children]))
+    def compute_height(self):
+        self.h = 0
+        for child in self.children:
+            self.h += child.h
 
     def recurse(self, node):
-        if isinstance(node, Text):
+        if isinstance(node, TextNode):
             self.text(node)
+        elif node.tag == "br":
+            self.flush()
+        elif node.tag == "input":
+            self.input(node)
+        elif node.tag == "button":
+            self.input(node)
         else:
-            if node.tag == "br":
-                self.new_line()
-            elif node.tag == "input" or node.tag == "button":
-                self.input(node)
-            else:
-                for child in node.children:
-                    self.recurse(child)
-
-    def new_line(self):
-        self.previous_word = None
-        self.cursor_x = self.x
-        last_line = self.children[-1] if self.children else None
-        new_line = LineLayout(self.node, self, last_line)
-        self.children.append(new_line)
+            for child in node.children:
+                self.recurse(child)
 
     def text(self, node):
-        weight = node.style["font-weight"]
-        style = node.style["font-style"]
-        size = float(node.style["font-size"][:-2])
-        font = get_font(size, weight, size)
         for word in node.text.split():
-            w = font.measureText(word)
-            if self.cursor_x + w > self.x + self.width:
-                self.new_line()
-            line = self.children[-1]
-            text = TextLayout(node, word, line, self.previous_word)
-            line.children.append(text)
-            self.previous_word = text
-            self.cursor_x += w + font.measureText(" ")
+            child = TextLayout(node, word)
+            child.size()
+            if self.children[-1].cx + child.w > self.w:
+                self.flush()
+            self.children[-1].append(child)
 
     def input(self, node):
-        w = INPUT_WIDTH_PX
-        if self.cursor_x + w > self.x + self.width:
-            self.new_line()
-        line = self.children[-1]
-        input = InputLayout(node, line, self.previous_word)
-        line.children.append(input)
-        self.previous_word = input
-        weight = node.style["font-weight"]
-        style = node.style["font-style"]
-        size = float(node.style["font-size"][:-2])
-        font = get_font(size, weight, size)
-        self.cursor_x += w + font.measureText(" ")
+        child = InputLayout(node)
+        child.size()
+        if self.children[-1].cx + child.w > self.w:
+            self.flush()
+        self.children[-1].append(child)
 
-    def paint(self, display_list):
-        cmds = []
+    def flush(self):
+        child = self.children[-1]
+        child.size()
+        self.children.append(LineLayout(self.node, self))
 
-        rect = skia.Rect.MakeLTRB(
-            self.x, self.y, self.x + self.width,
-            self.y + self.height)
-
-        bgcolor = self.node.style.get("background-color",
-                                 "transparent")
-        if bgcolor != "transparent":
-            radius = float(self.node.style.get("border-radius", "0px")[:-2])
-            cmds.append(DrawRRect(rect, radius, bgcolor))
- 
+    def position(self):
+        cy = self.y
         for child in self.children:
-            child.paint(cmds)
+            child.x = self.x
+            child.y = cy
+            child.position()
+            cy += child.h
 
-        cmds = paint_visual_effects(self.node, cmds, rect)
-        display_list.extend(cmds)
+    def draw(self, to):
+        for child in self.children:
+            child.draw(to)
 
-    def __repr__(self):
-        return "InlineLayout(x={}, y={}, width={}, height={})".format(
-            self.x, self.y, self.width, self.height)
+def px(s):
+    if s.endswith("px"):
+        return int(s[:-2])
+    else:
+        return 0
+
+class BlockLayout:
+    def __init__(self, node, parent):
+        self.node = node
+        self.parent = parent
+
+        self.x = -1
+        self.y = -1
+        self.w = -1
+        self.h = -1
+
+    def has_block_children(self):
+        for child in self.node.children:
+            if isinstance(child, TextNode):
+                if not child.text.isspace():
+                    return False
+            elif child.style.get("display", "block") == "inline":
+                return False
+        return True
+
+    def size(self):
+        self.children = []
+        # block layout here
+        if self.has_block_children():
+            for child in self.node.children:
+                if isinstance(child, TextNode): continue
+                self.children.append(BlockLayout(child, self))
+        else:
+            self.children.append(InlineLayout(self.node, self))
+
+        self.mt = px(self.node.style.get("margin-top", "0px"))
+        self.bt = px(self.node.style.get("border-top-width", "0px"))
+        self.pt = px(self.node.style.get("padding-top", "0px"))
+        self.mr = px(self.node.style.get("margin-right", "0px"))
+        self.br = px(self.node.style.get("border-right-width", "0px"))
+        self.pr = px(self.node.style.get("padding-right", "0px"))
+        self.mb = px(self.node.style.get("margin-bottom", "0px"))
+        self.bb = px(self.node.style.get("border-bottom-width", "0px"))
+        self.pb = px(self.node.style.get("padding-bottom", "0px"))
+        self.ml = px(self.node.style.get("margin-left", "0px"))
+        self.bl = px(self.node.style.get("border-left-width", "0px"))
+        self.pl = px(self.node.style.get("padding-left", "0px"))
+
+        self.w = self.parent.w - self.parent.pl - self.parent.pr \
+            - self.parent.bl - self.parent.br \
+            - self.ml - self.mr
+        for child in self.children:
+            child.size()
+        self.compute_height()
+
+    def compute_height(self):
+        self.h = self.pt + self.pb
+        for child in self.children:
+            self.h += child.mt + child.h + child.mb
+
+    def position(self):
+        self.y += self.mt
+        self.x += self.ml
+
+        y = self.y + self.pt
+        for child in self.children:
+            child.x = self.x + self.pl + self.bl
+            child.y = y
+            child.position()
+            y += child.mt + child.h + child.mb
+
+    def draw(self, to):
+        if self.node.tag == "pre":
+            x2, y2 = self.x + self.w, self.y + self.h
+            to.append(DrawRect(self.x, self.y, x2, y2, "gray"))
+        for child in self.children:
+            child.draw(to)
 
 class DocumentLayout:
     def __init__(self, node):
         self.node = node
         self.parent = None
-        self.previous = None
         self.children = []
 
-    def layout(self):
-        child = BlockLayout(self.node, self, None)
+        self.x = -1
+        self.y = -1
+        self.w = -1
+        self.h = -1
+
+    def size(self):
+        child = BlockLayout(self.node, self)
         self.children.append(child)
 
-        self.width = WIDTH - 2*HSTEP
-        self.x = HSTEP
-        self.y = VSTEP
-        child.layout()
-        self.height = child.height + 2*VSTEP
+        self.w = WIDTH
+        self.mt = self.bt = self.pt = 0
+        self.mr = self.br = self.pr = 0
+        self.mb = self.bb = self.pb = 0
+        self.ml = self.bl = self.pl = 0
 
-    def paint(self, display_list):
-        self.children[0].paint(display_list)
+        child.size()
+        self.compute_height()
 
-    def __repr__(self):
-        return "DocumentLayout()"
+    def compute_height(self):
+        self.h = self.children[0].h
 
-INPUT_WIDTH_PX = 200
+    def position(self):
+        child = self.children[0]
+        child.x = self.x = 0
+        child.y = self.y = 0
+        child.position()
 
-class LineLayout:
-    def __init__(self, node, parent, previous):
-        self.node = node
-        self.parent = parent
-        self.previous = previous
-        self.children = []
-        self.x = None
-        self.y = None
-        self.width = None
-        self.height = None
+    def draw(self, to):
+        self.children[0].draw(to)
 
-    def layout(self):
-        self.width = self.parent.width
-        self.x = self.parent.x
+class DrawText:
+    def __init__(self, x1, y1, text, font, color):
+        self.x1 = x1
+        self.y1 = y1
+        self.text = text
+        self.font = font
+        self.color = color
 
-        if self.previous:
-            self.y = self.previous.y + self.previous.height
+        self.y2 = y1 + font.measure("linespace")
+
+    def draw(self, scroll, canvas):
+        canvas.create_text(
+            self.x1, self.y1 - scroll,
+            text=self.text,
+            font=self.font,
+            fill=self.color,
+            anchor='nw',
+        )
+
+class DrawRect:
+    def __init__(self, x1, y1, x2, y2, color):
+        self.x1 = x1
+        self.y1 = y1
+        self.x2 = x2
+        self.y2 = y2
+        self.color = color
+
+    def draw(self, scroll, canvas):
+        canvas.create_rectangle(
+            self.x1, self.y1 - scroll,
+            self.x2, self.y2 - scroll,
+            width=0,
+            fill=self.color,
+        )
+
+def find_links(node, lst):
+    if not isinstance(node, ElementNode): return
+    if node.tag == "link" and \
+       node.attributes.get("rel", "") == "stylesheet" and \
+       "href" in node.attributes:
+        lst.append(node.attributes["href"])
+    for child in node.children:
+        find_links(child, lst)
+    return lst
+
+def resolve_url(url, current):
+    if "://" in url:
+        return url
+    elif url.startswith("/"):
+        return "/".join(current.split("/")[:3]) + url
+    else:
+        return current.rsplit("/", 1)[0] + "/" + url
+
+def find_layout(x, y, tree):
+    for child in reversed(tree.children):
+        result = find_layout(x, y, child)
+        if result: return result
+    if tree.x <= x < tree.x + tree.w and \
+       tree.y <= y < tree.y + tree.h:
+        return tree
+
+def find_inputs(elt, out):
+    if not isinstance(elt, ElementNode): return
+    if elt.tag == "input" and "name" in elt.attributes:
+        out.append(elt)
+    for child in elt.children:
+        find_inputs(child, out)
+    return out
+
+def find_scripts(node, out):
+    if not isinstance(node, ElementNode): return
+    if node.tag == "script" and \
+       "src" in node.attributes:
+        out.append(node.attributes["src"])
+    for child in node.children:
+        find_scripts(child, out)
+    return out
+
+def find_selected(node, sel, out):
+    if not isinstance(node, ElementNode): return
+    if sel.matches(node):
+        out.append(node)
+    for child in node.children:
+        find_selected(child, sel, out)
+    return out
+
+def layout_for_node(tree, node):
+    if tree.node == node:
+        return tree
+    for child in tree.children:
+        out = layout_for_node(child, node)
+        if out: return out
+
+def is_link(node):
+    return isinstance(node, ElementNode) \
+        and node.tag == "a" and "href" in node.attributes
+
+def drawTree(node, indent=0):
+    print(" "*indent, type(node).__name__, " ", node.node, sep="")
+    for child in node.children:
+        drawTree(child, indent + 2)
+
+class Browser:
+    def __init__(self):
+        self.window = tkinter.Tk()
+        self.canvas = tkinter.Canvas(
+            self.window,
+            width=WIDTH,
+            height=HEIGHT
+        )
+        self.canvas.pack()
+
+        self.history = []
+        self.focus = None
+        self.address_bar = ""
+        self.scroll = 0
+
+        self.timer = Timer()
+        self.window.bind("<Down>", self.scrolldown)
+        self.window.bind("<Button-1>", self.handle_click)
+        self.window.bind("<Key>", self.keypress)
+        self.window.bind("<Return>", self.pressenter)
+        self.display_list = []
+
+    def handle_click(self, e):
+        self.focus = None
+        if e.y < 60: # Browser chrome
+            if 10 <= e.x < 35 and 10 <= e.y < 50:
+                self.go_back()
+            elif 50 <= e.x < 790 and 10 <= e.y < 50:
+                self.focus = "address bar"
+                self.address_bar = ""
+                self.render()
         else:
-            self.y = self.parent.y
+            x, y = e.x, e.y + self.scroll - 60
+            obj = find_layout(x, y, self.document)
+            if not obj: return
+            elt = obj.node
+            if elt and self.dispatch_event("click", elt): return
+            while elt:
+                if isinstance(elt, TextNode):
+                    pass
+                elif is_link(elt):
+                    url = resolve_url(elt.attributes["href"], self.url)
+                    return self.load(url)
+                elif elt.tag == "input":
+                    elt.attributes["value"] = ""
+                    self.focus = obj
+                    return self.reflow(self.focus)
+                elif elt.tag == "button":
+                    self.submit_form(elt)
+                elt = elt.parent
 
-        for word in self.children:
-            word.layout()
+    def keypress(self, e):
+        if len(e.char) == 0: return
+        if not (0x20 <= ord(e.char) < 0x7f): return
 
-        if not self.children:
-            self.height = 0
+        if not self.focus:
             return
-
-        max_ascent = max([-word.font.getMetrics().fAscent 
-                          for word in self.children])
-        baseline = self.y + 1.25 * max_ascent
-        for word in self.children:
-            word.y = baseline + word.font.getMetrics().fAscent
-        max_descent = max([word.font.getMetrics().fDescent
-                           for word in self.children])
-        self.height = 1.25 * (max_ascent + max_descent)
-
-    def paint(self, display_list):
-        for child in self.children:
-            child.paint(display_list)
-
-    def __repr__(self):
-        return "LineLayout(x={}, y={}, width={}, height={})".format(
-            self.x, self.y, self.width, self.height)
-
-class TextLayout:
-    def __init__(self, node, word, parent, previous):
-        self.node = node
-        self.word = word
-        self.children = []
-        self.parent = parent
-        self.previous = previous
-        self.x = None
-        self.y = None
-        self.width = None
-        self.height = None
-        self.font = None
-
-    def layout(self):
-        weight = self.node.style["font-weight"]
-        style = self.node.style["font-style"]
-        if style == "normal": style = "roman"
-        size = float(self.node.style["font-size"][:-2])
-        self.font = get_font(size, weight, style)
-
-        # Do not set self.y!!!
-        self.width = self.font.measureText(self.word)
-
-        if self.previous:
-            space = self.previous.font.measureText(" ")
-            self.x = self.previous.x + space + self.previous.width
+        elif self.focus == "address bar":
+            self.address_bar += e.char
+            self.render()
         else:
-            self.x = self.parent.x
+            self.focus.node.attributes["value"] += e.char
+            self.dispatch_event("change", self.focus.node)
+            self.reflow(self.focus)
 
-        self.height = linespace(self.font)
+    def submit_form(self, elt):
+        while elt and elt.tag != "form":
+            elt = elt.parent
+        if not elt: return
+        if self.dispatch_event("submit", elt): return
+        inputs = find_inputs(elt, [])
+        body = ""
+        for input in inputs:
+            name = input.attributes["name"]
+            value = input.attributes.get("value", "")
+            body += "&" + name + "=" + value.replace(" ", "%20")
+        body = body[1:]
+        url = resolve_url(elt.attributes["action"], self.url)
+        self.load(url, body)
 
-    def paint(self, display_list):
-        color = self.node.style["color"]
-        display_list.append(
-            DrawText(self.x, self.y, self.word, self.font, color))
-    
-    def __repr__(self):
-        return "TextLayout(x={}, y={}, width={}, height={}".format(
-            self.x, self.y, self.width, self.height)
+    def pressenter(self, e):
+        if self.focus == "address bar":
+            self.focus = None
+            self.load(self.address_bar)
 
-class InputLayout:
-    def __init__(self, node, parent, previous):
-        self.node = node
-        self.children = []
-        self.parent = parent
-        self.previous = previous
-        self.x = None
-        self.y = None
-        self.width = None
-        self.height = None
-        self.font = None
+    def go_back(self):
+        if len(self.history) > 1:
+            self.history.pop()
+            back = self.history.pop()
+            self.load(back)
 
-    def layout(self):
-        weight = self.node.style["font-weight"]
-        style = self.node.style["font-style"]
-        if style == "normal": style = "roman"
-        size = float(self.node.style["font-size"][:-2])
-        self.font = get_font(size, weight, style)
+    def load(self, url, body=None):
+        self.timer.start("Downloading")
+        self.address_bar = url
+        self.url = url
+        self.history.append(url)
+        header, body = request(url, payload=body)
+        self.timer.start("Parsing HTML")
+        self.nodes = parse(lex(body))
+        
+        self.timer.start("Parsing CSS")
+        with open("browser8.css") as f:
+            self.rules = CSSParser(f.read()).parse()
 
-        self.width = style_length(
-            self.node, "width", INPUT_WIDTH_PX)
-        self.height = style_length(
-            self.node, "height", linespace(self.font))
+        for link in find_links(self.nodes, []):
+            header, body = request(resolve_url(link, url))
+            self.rules.extend(CSSParser(body).parse())
 
-        if self.previous:
-            space = self.previous.font.measureText(" ")
-            self.x = self.previous.x + space + self.previous.width
-        else:
-            self.x = self.parent.x
+        self.rules.sort(key=lambda x: x[0].priority())
+        self.rules.reverse()
 
-    def paint(self, display_list):
-        cmds = []
+        self.timer.start("Running JS")
+        self.setup_js()
+        for script in find_scripts(self.nodes, []):
+            header, body = request(resolve_url(script, self.history[-1]))
+            try:
+                print("Script returned: ", self.js.evaljs(body))
+            except dukpy.JSRuntimeError as e:
+                print("Script", script, "crashed", e)
+        self.layout(self.nodes)
 
-        rect = skia.Rect.MakeLTRB(
-            self.x, self.y, self.x + self.width,
-            self.y + self.height)
-
-        bgcolor = self.node.style.get("background-color",
-                                 "transparent")
-        if bgcolor != "transparent":
-            radius = float(self.node.style.get("border-radius", "0px")[:-2])
-            cmds.append(DrawRRect(rect, radius, bgcolor))
-
-        if self.node.tag == "input":
-            text = self.node.attributes.get("value", "")
-        elif self.node.tag == "button":
-            text = self.node.children[0].text
-
-        color = self.node.style["color"]
-        cmds.append(DrawText(self.x, self.y,
-                             text, self.font, color))
-
-        cmds = paint_visual_effects(self.node, cmds, rect)
-        display_list.extend(cmds)
-
-    def __repr__(self):
-        return "InputLayout(x={}, y={}, width={}, height={})".format(
-            self.x, self.y, self.width, self.height)
-
-def style_length(node, style_name, default_value):
-    style_val = node.style.get(style_name)
-    if style_val:
-        return int(style_val[:-2])
-    else:
-        return default_value
-
-def paint_visual_effects(node, cmds, rect):
-    opacity = float(node.style.get("opacity", "1.0"))
-    blend_mode = parse_blend_mode(node.style.get("mix-blend-mode"))
-    (translation, rotation) = parse_transform(node.style.get("transform", ""))
-
-    border_radius = float(node.style.get("border-radius", "0px")[:-2])
-    if node.style.get("overflow", "visible") == "clip":
-        clip_radius = border_radius
-    else:
-        clip_radius = 0
-
-    needs_clip = node.style.get("overflow", "visible") == "clip"
-    needs_blend_isolation = blend_mode != skia.BlendMode.kSrcOver or \
-        needs_clip or opacity != 1.0
-
-    return [
-        Transform(translation, rotation, rect, [
-            SaveLayer(skia.Paint(BlendMode=blend_mode, Alphaf=opacity), [
-                ClipRRect(rect, clip_radius,
-                    cmds,
-                should_clip=needs_clip),
-            ], should_save=needs_blend_isolation),
-        ],
-        should_transform=translation != None or rotation != None)
-    ]
-
-XHR_ONLOAD_CODE = "__runXHROnload(dukpy.out, dukpy.handle)"
-
-class JSContext:
-    def __init__(self, tab):
-        self.tab = tab
-
-        self.interp = dukpy.JSInterpreter()
-        self.interp.export_function("log", print)
-        self.interp.export_function("querySelectorAll",
-            self.querySelectorAll)
-        self.interp.export_function("getAttribute",
-            self.getAttribute)
-        self.interp.export_function("innerHTML_set", self.innerHTML_set)
-        self.interp.export_function("style_set", self.style_set)
-        self.interp.export_function("XMLHttpRequest_send",
-            self.XMLHttpRequest_send)
-        self.interp.export_function("now",
-            self.now)
-        self.interp.export_function("requestAnimationFrame",
-            self.requestAnimationFrame)
-        with open("runtime14.js") as f:
-            self.interp.evaljs(f.read())
-
+    def setup_js(self):
+        self.js = dukpy.JSInterpreter()
         self.node_to_handle = {}
         self.handle_to_node = {}
+        self.js.export_function("log", print)
+        self.js.export_function("querySelectorAll", self.js_querySelectorAll)
+        self.js.export_function("getAttribute", self.js_getAttribute)
+        self.js.export_function("innerHTML", self.js_innerHTML)
+        with open("runtime9.js") as f:
+            self.js.evaljs(f.read())
 
-    def run(self, script, code):
+    def js_querySelectorAll(self, sel):
+        selector, _ = CSSParser(sel + "{").selector(0)
+        elts = find_selected(self.nodes, selector, [])
+        return [self.make_handle(elt) for elt in elts]
+    
+    def js_getAttribute(self, handle, attr):
+        elt = self.handle_to_node[handle]
+        return elt.attributes.get(attr, None)
+
+    def js_innerHTML(self, handle, s):
         try:
-            print("Script returned: ", self.interp.evaljs(code))
-        except dukpy.JSRuntimeError as e:
-            print("Script", script, "crashed", e)
+            doc = parse(lex("<!doctype><html><body>" + s + "</body></html>"))
+            new_nodes = doc.children[0].children
+            elt = self.handle_to_node[handle]
+            elt.children = new_nodes
+            for child in elt.children:
+                child.parent = elt
+            self.reflow(layout_for_node(self.document, elt))
+        except:
+            import traceback
+            traceback.print_exc()
+            raise
 
     def dispatch_event(self, type, elt):
-        handle = self.node_to_handle.get(elt, -1)
-        do_default = self.interp.evaljs(
-            EVENT_DISPATCH_CODE, type=type, handle=handle)
-        return not do_default
+       handle = self.node_to_handle.get(elt, -1)
+       do_default = self.js.evaljs("__runHandlers({}, \"{}\")".format(handle, type))
+       return not do_default
 
-    def get_handle(self, elt):
+    def make_handle(self, elt):
         if elt not in self.node_to_handle:
             handle = len(self.node_to_handle)
             self.node_to_handle[elt] = handle
@@ -841,1048 +859,59 @@ class JSContext:
             handle = self.node_to_handle[elt]
         return handle
 
-    def querySelectorAll(self, selector_text):
-        selector = CSSParser(selector_text).selector()
-        nodes = [node for node
-                 in tree_to_list(self.tab.nodes, [])
-                 if selector.matches(node)]
-        return [self.get_handle(node) for node in nodes]
+    def layout(self, tree):
+        self.document = DocumentLayout(tree)
+        self.reflow(self.document)
 
-    def getAttribute(self, handle, attr):
-        elt = self.handle_to_node[handle]
-        return elt.attributes.get(attr, None)
-
-    def innerHTML_set(self, handle, s):
-        doc = HTMLParser(
-            "<html><body>" + s + "</body></html>").parse()
-        new_nodes = doc.children[0].children
-        elt = self.handle_to_node[handle]
-        elt.children = new_nodes
-        for child in elt.children:
-            child.parent = elt
-        self.tab.set_needs_pipeline_update()
-
-    def style_set(self, handle, s):
-        elt = self.handle_to_node[handle]
-        elt.attributes["style"] = s;
-        self.tab.set_needs_pipeline_update()
-
-    def xhr_onload(self, out, handle):
-        do_default = self.interp.evaljs(
-            XHR_ONLOAD_CODE, out=out, handle=handle)
-
-    def XMLHttpRequest_send(
-        self, method, url, body, is_async, handle):
-        full_url = resolve_url(url, self.tab.url)
-        if not self.tab.allowed_request(full_url):
-            raise Exception("Cross-origin XHR blocked by CSP")
-
-        def run_load():
-            headers, out = request(
-                full_url, self.tab.url, payload=body)
-            handle_local = handle
-            if url_origin(full_url) != url_origin(self.tab.url):
-                raise Exception(
-                    "Cross-origin XHR request not allowed")
-            self.tab.main_thread_runner.schedule_script_task(
-                Task(self.xhr_onload, out, handle_local))
-            return out
-
-        if not is_async:
-            run_load(is_async)
-        else:
-            load_thread = threading.Thread(target=run_load, args=())
-            load_thread.start()
-
-    def now(self):
-        return int(time.time() * 1000)
-
-    def requestAnimationFrame(self):
-        self.tab.request_animation_frame_callback()
-
-SCROLL_STEP = 100
-CHROME_PX = 100
-
-USE_BROWSER_THREAD = True
-
-def set_timeout(func, sec):
-    t = threading.Timer(sec, func)
-    t.start()
-
-def clamp_scroll(scroll, tab_height):
-    return max(0, min(scroll, tab_height - (HEIGHT - CHROME_PX)))
-
-def animate_style(node, old_style, new_style, tab):
-    if not old_style:
-        return
-    if not "transition" in old_style or not "transition" in new_style:
-        return
-    if not old_style["transition"] == "opacity" or not new_style["transition"] == "opacity":
-        return
-    if old_style["opacity"] == new_style["opacity"]:
-        return
-    tab.animations[node] = start_opacity_animation(float(old_style["opacity"]), new_style, tab)
-
-ANIMATION_FRAME_COUNT = 60
-
-def start_opacity_animation(old_opacity, new_style, tab):
-    new_opacity = float(new_style["opacity"])
-    change_per_frame = (new_opacity - old_opacity) / ANIMATION_FRAME_COUNT
-    return Animation("opacity", old_opacity, change_per_frame, new_style, tab)
-
-def style(node, rules, tab):
-    old_style = None
-    if hasattr(node, 'style'):
-        old_style = node.style
-
-    node.style = {}
-    for property, default_value in INHERITED_PROPERTIES.items():
-        if node.parent:
-            node.style[property] = node.parent.style[property]
-        else:
-            node.style[property] = default_value
-    for selector, body in rules:
-        if not selector.matches(node): continue
-        for property, value in body.items():
-            computed_value = compute_style(node, property, value)
-            if not computed_value: continue
-            node.style[property] = computed_value
-    if isinstance(node, Element) and "style" in node.attributes:
-        pairs = CSSParser(node.attributes["style"]).body()
-        for property, value in pairs.items():
-            computed_value = compute_style(node, property, value)
-            node.style[property] = computed_value
-    animate_style(node, old_style, node.style, tab)
-    for child in node.children:
-        style(child, rules, tab)
-
-class Animation:
-    def __init__(
-        self, property_name, old_value, change_per_frame, computed_style, tab):
-        self.property_name = property_name
-        self.old_value = old_value
-        self.change_per_frame = change_per_frame
-        self.computed_style = computed_style
-        self.tab = tab
-        self.frame_count = 0
-        tab.set_needs_animation_frame()
-
-    def animate(self):
-        self.frame_count += 1
-        self.computed_style[self.property_name] = \
-            self.old_value + self.change_per_frame * self.frame_count
-        needs_another_frame = self.frame_count < ANIMATION_FRAME_COUNT
-        if needs_another_frame:
-            self.tab.set_needs_paint()
-            self.tab.set_needs_animation_frame()
-        return needs_another_frame
-
-class CompositedLayer:
-    def __init__(self, bounds=None, first_chunk=None):
-        self.surface = None
-        self.chunks = []
-        self.first_chunk = first_chunk
-        if first_chunk:
-            self.chunks.append(first_chunk)
-
-    def can_merge(self, chunk):
-        if len(self.chunks) == 0:
-            return not chunk.needs_compositing()
-        return  \
-            self.chunks[0].composited_item() == chunk.composited_item()
-
-    def bounds(self):
-        retval = skia.Rect.MakeEmpty()
-        for chunk in self.chunks:
-            retval.join(chunk.bounds())
-        return retval
-
-    def append(self, chunk):
-        assert self.can_merge(chunk)
-        self.chunks.append(chunk)
-
-    def overlaps(self, rect):
-        return skia.Rect.Intersects(self.bounds(), rect)
-
-    def draw(self, canvas, draw_offset):
-        if not self.surface:
-            return
-        (offset_x, offset_y) = draw_offset
-        bounds = self.bounds()
-        offset_x += bounds.left()
-        offset_y += bounds.top()
-
-        def op():
-            self.surface.draw(canvas, 0, 0)
-
-        canvas.save()
-        canvas.translate(offset_x, offset_y)
-
-        if self.first_chunk:
-            self.first_chunk.draw(canvas, op, False)
-        else:
-            self.surface.draw(canvas, 0, 0)
-        canvas.restore()
-
-    def raster(self):
-        bounds = self.bounds()
-        if bounds.isEmpty():
-            return
-        irect = bounds.roundOut()
-        if not self.surface:
-            self.surface = skia.Surface(irect.width(), irect.height())
-        canvas = self.surface.getCanvas()
-        canvas.clear(skia.ColorWHITE)
-        canvas.save()
-        canvas.translate(-bounds.left(), -bounds.top())
-        for chunk in self.chunks:
-            chunk.raster(canvas)
-        canvas.restore()
-
-class Tab:
-    def __init__(self, commit_func):
-        self.history = []
-        self.focus = None
-        self.url = None
-        self.scroll = 0
-        self.scroll_changed_in_tab = False
-        self.needs_raf_callbacks = False
-        self.display_scheduled = False
-        self.needs_pipeline_update = False
-        self.needs_paint = False
-        self.commit_func = commit_func
-        if USE_BROWSER_THREAD:
-            self.main_thread_runner = MainThreadRunner(self)
-        else:
-            self.main_thread_runner = SingleThreadedTaskRunner(self)
-        self.main_thread_runner.start()
-
-        self.time_in_style_layout_and_paint = 0.0
-        self.animations = {}
+    def reflow(self, obj):
+        self.timer.start("Style")
+        style(obj.node, obj.node.parent, self.rules)
+        self.timer.start("Layout (phase 1A)")
+        obj.size()
+        self.timer.start("Layout (phase 1B)")
+        while obj.parent:
+            obj.parent.compute_height()
+            obj = obj.parent
+        self.timer.start("Layout (phase 2)")
+        self.document.position()
+        self.timer.start("Display list")
         self.display_list = []
-
-        with open("browser8.css") as f:
-            self.default_style_sheet = CSSParser(f.read()).parse()
-
-    def allowed_request(self, url):
-        return self.allowed_origins == None or \
-            url_origin(url) in self.allowed_origins
-
-    def cookie_string(self):
-        origin = url_origin(self.history[-1])
-        cookie_string = ""
-        if not origin in self.cookies:
-            return cookie_string
-        for key, value in self.cookies[origin].items():
-            cookie_string += "&" + key + "=" + value
-        return cookie_string[1:]
-
-    def script_run_wrapper(self, script, script_text):
-        return Task(self.js.run, script, script_text)
-
-    def load(self, url, body=None):
-        self.main_thread_runner.clear_pending_tasks()
-        headers, body = request(url, self.url, payload=body)
-        self.scroll = 0
-        self.scroll_changed_in_tab = True
-        self.url = url
-        self.history.append(url)
-
-        self.allowed_origins = None
-        if "content-security-policy" in headers:
-           csp = headers["content-security-policy"].split()
-           if len(csp) > 0 and csp[0] == "default-src":
-               self.allowed_origins = csp[1:]
-
-        self.nodes = HTMLParser(body).parse()
-
-        self.js = JSContext(self)
-        scripts = [node.attributes["src"] for node
-                   in tree_to_list(self.nodes, [])
-                   if isinstance(node, Element)
-                   and node.tag == "script"
-                   and "src" in node.attributes]
-
-        async_requests = []
-        script_results = {}
-        for script in scripts:
-            script_url = resolve_url(script, url)
-            if not self.allowed_request(script_url):
-                print("Blocked script", script, "due to CSP")
-                continue
-            async_requests.append({
-                "url": script_url,
-                "type": "script",
-                "thread": async_request(
-                    script_url, url, script_results)
-            })
- 
-        self.rules = self.default_style_sheet.copy()
-        links = [node.attributes["href"]
-                 for node in tree_to_list(self.nodes, [])
-                 if isinstance(node, Element)
-                 and node.tag == "link"
-                 and "href" in node.attributes
-                 and node.attributes.get("rel") == "stylesheet"]
-
-        style_results = {}
-        for link in links:
-            style_url = resolve_url(link, url)
-            if not self.allowed_request(style_url):
-                print("Blocked style", link, "due to CSP")
-                continue
-            async_requests.append({
-                "url": style_url,
-                "type": "style sheet",
-                "thread": async_request(style_url, url, style_results)
-            })
-
-        for async_req in async_requests:
-            async_req["thread"].join()
-            req_url = async_req["url"]
-            if async_req["type"] == "script":
-                self.main_thread_runner.schedule_script_task(
-                    Task(self.js.run, req_url,
-                        script_results[req_url]['body']))
-            else:
-                self.rules.extend(
-                    CSSParser(
-                        style_results[req_url]['body']).parse())
-
-        self.set_needs_pipeline_update()
-
-    def apply_scroll(self, scroll):
-        self.scroll = scroll
-
-    def set_needs_paint(self):
-        self.needs_paint = True
-
-    def set_needs_pipeline_update(self):
-        self.needs_pipeline_update = True
-        self.needs_paint = True
-        self.set_needs_animation_frame()
-
-    def set_needs_animation_frame(self):
-        def callback():
-            self.display_scheduled = False
-            self.main_thread_runner.schedule_animation_frame()
-        if not self.display_scheduled:
-            if USE_BROWSER_THREAD:
-                set_timeout(callback, REFRESH_RATE_SEC)
-            self.display_scheduled = True
-
-    def request_animation_frame_callback(self):
-        self.needs_raf_callbacks = True
-        self.set_needs_animation_frame()
-
-    def compute_document_bounds(self):
-#        print('compute_document_bounds')
-        rect = skia.Rect.MakeEmpty()
-        for display_item in self.display_list:
-#            print("   " + str(display_item) + " " + str(display_item.bounds()))
-            rect.join(display_item.bounds())
-        return rect
-
-    def print_display_list(self):
-        print("Display list:")
-        out = ""
-        for display_item in self.display_list:
-            out += display_item.repr_recursive(indent=2, include_noop=True)
-        print(out)
-
-    def run_animation_frame(self):
-        if self.needs_raf_callbacks:
-            self.needs_raf_callbacks = False
-            self.js.interp.evaljs("__runRAFHandlers()")
-
-        self.run_rendering_pipeline()
-
-#        self.print_display_list()
-        document_bounds = self.compute_document_bounds()
-        clamped_scroll = clamp_scroll(self.scroll,
-            document_bounds.height())
-        if clamped_scroll != self.scroll:
-            self.scroll_changed_in_tab = True
-        self.scroll = clamped_scroll
-
-        self.commit_func(
-            self.url, clamped_scroll if self.scroll_changed_in_tab \
-                else None, 
-            document_bounds,
-            self.display_list)
-        self.scroll_changed_in_tab = False
-
-    def run_rendering_pipeline(self):
-        timer = None
-
-        to_delete = []
-        for animation_key in self.animations:
-            if not self.animations[animation_key].animate():
-                to_delete.append(animation_key)
-
-        for key in to_delete:
-            del self.animations[key]
-
-        if self.needs_pipeline_update:
-            timer = Timer()
-            timer.start()
-
-            style(self.nodes, sorted(self.rules,
-                key=cascade_priority), self)
-            self.document = DocumentLayout(self.nodes)
-            self.document.layout()
-
-        if self.needs_paint:
-            self.display_list = []
-            self.document.paint(self.display_list)
-            if self.focus:
-                obj = [obj for obj in tree_to_list(self.document, [])
-                       if obj.node == self.focus][0]
-                text = self.focus.attributes.get("value", "")
-                x = obj.x + obj.font.measureText(text)
-                y = obj.y
-                self.display_list.append(
-                    DrawLine(x, y, x, y + obj.height))
-            self.needs_paint = False
-
-        self.needs_pipeline_update = False
-
-        if timer:
-            self.time_in_style_layout_and_paint += timer.stop()
-
-    def click(self, x, y):
-        self.run_rendering_pipeline()
-        self.focus = None
-        y += self.scroll
-        objs = [obj for obj in tree_to_list(self.document, [])
-                if obj.x <= x < obj.x + obj.width
-                and obj.y <= y < obj.y + obj.height]
-        if not objs: return
-        elt = objs[-1].node
-        if elt and self.js.dispatch_event("click", elt): return
-        while elt:
-            if isinstance(elt, Text):
-                pass
-            elif elt.tag == "a" and "href" in elt.attributes:
-                url = resolve_url(elt.attributes["href"], self.url)
-                self.load(url)
-                return
-            elif elt.tag == "input":
-                elt.attributes["value"] = ""
-                if elt != self.focus:
-                    self.set_needs_pipeline_update()
-                self.focus = elt
-                return
-            elif elt.tag == "button":
-                while elt:
-                    if elt.tag == "form" and "action" in elt.attributes:
-                        return self.submit_form(elt)
-                    elt = elt.parent
-            elt = elt.parent
-
-    def submit_form(self, elt):
-        if self.js.dispatch_event("submit", elt): return
-        inputs = [node for node in tree_to_list(elt, [])
-                  if isinstance(node, Element)
-                  and node.tag == "input"
-                  and "name" in node.attributes]
-
-        body = ""
-        for input in inputs:
-            name = input.attributes["name"]
-            value = input.attributes.get("value", "")
-            name = urllib.parse.quote(name)
-            value = urllib.parse.quote(value)
-            body += "&" + name + "=" + value
-        body = body [1:]
-
-        url = resolve_url(elt.attributes["action"], self.url)
-        self.load(url, body)
-
-
-    def keypress(self, char):
-        if self.focus:
-            if self.js.dispatch_event("keydown", self.focus): return
-            self.focus.attributes["value"] += char
-            self.set_needs_pipeline_update()
-
-    def go_back(self):
-        if len(self.history) > 1:
-            self.history.pop()
-            back = self.history.pop()
-            self.load(back)
-
-WIDTH, HEIGHT = 800, 600
-HSTEP, VSTEP = 13, 18
-
-class Task:
-    def __init__(self, task_code, *args):
-        self.task_code = task_code
-        self.args = args
-        self.__name__ = "task"
-
-    def __call__(self):
-        self.task_code(*self.args)
-        self.task_code = None
-        self.args = None
-
-class TaskQueue:
-    def __init__(self):
-        self.tasks = []
-
-    def add_task(self, task_code):
-        self.tasks.append(task_code)
-
-    def has_tasks(self):
-        retval = len(self.tasks) > 0
-        return retval
-
-    def get_next_task(self):
-        retval = self.tasks.pop(0)
-        return retval
-
-    def clear(self):
-        self.tasks = []
-
-class SingleThreadedTaskRunner:
-    def __init__(self, tab):
-        self.tab = tab
-
-    def schedule_scroll(self, scroll):
-        self.tab.apply_scroll(scroll)
-
-    def schedule_animation_frame(self):
-        self.tab.run_animation_frame()
-
-    def schedule_script_task(self, script):
-        script()
-
-    def schedule_browser_task(self, callback):
-        callback()
-
-    def schedule_scroll(self, scroll):
-        self.tab.scroll = scroll
-
-    def clear_pending_tasks(self):
-        pass
-
-    def start(self):    
-        pass
-
-    def set_needs_quit(self):
-        pass
-
-    def run(self):
-        pass
-
-class MainThreadRunner:
-    def __init__(self, tab):
-        self.lock = threading.Lock()
-        self.condition = threading.Condition(self.lock)
-        self.tab = tab
-        self.needs_animation_frame = False
-        self.main_thread = threading.Thread(target=self.run, args=())
-        self.script_tasks = TaskQueue()
-        self.browser_tasks = TaskQueue()
-        self.needs_quit = False
-        self.pending_scroll = None
-
-    def schedule_animation_frame(self):
-        self.lock.acquire(blocking=True)
-        self.needs_animation_frame = True
-        self.condition.notify_all()
-        self.lock.release()
-
-    def schedule_script_task(self, script):
-        self.lock.acquire(blocking=True)
-        self.script_tasks.add_task(script)
-        self.condition.notify_all()
-        self.lock.release()
-
-    def schedule_browser_task(self, callback):
-        self.lock.acquire(blocking=True)
-        self.browser_tasks.add_task(callback)
-        self.condition.notify_all()
-        self.lock.release()
-
-    def set_needs_quit(self):
-        self.lock.acquire(blocking=True)
-        self.needs_quit = True
-        self.condition.notify_all()
-        self.lock.release()
-
-    def schedule_scroll(self, scroll):
-        self.lock.acquire(blocking=True)
-        self.pending_scroll = scroll
-        self.condition.notify_all()
-        self.lock.release()
-
-    def clear_pending_tasks(self):
-        self.needs_animation_frame = False
-        self.script_tasks.clear()
-        self.browser_tasks.clear()
-        self.pending_scroll = None
-
-    def start(self):
-        self.main_thread.start()
-
-    def run(self):
-        while True:
-            if self.needs_quit:
-                return;
-
-            self.lock.acquire(blocking=True)
-            needs_animation_frame = self.needs_animation_frame
-            self.needs_animation_frame = False
-            pending_scroll = self.pending_scroll
-            self.pending_scroll = None
-            self.lock.release()
-            if pending_scroll:
-                self.tab.apply_scroll(pending_scroll)
-            if needs_animation_frame:
-                self.tab.run_animation_frame()
-
-            browser_method = None
-            self.lock.acquire(blocking=True)
-            if self.browser_tasks.has_tasks():
-                browser_method = self.browser_tasks.get_next_task()
-            self.lock.release()
-            if browser_method:
-                browser_method()
-
-            script = None
-            self.lock.acquire(blocking=True)
-            if self.script_tasks.has_tasks():
-                script = self.script_tasks.get_next_task()
-            self.lock.release()
-            if script:
-                script()
-
-            self.lock.acquire(blocking=True)
-            if not self.script_tasks.has_tasks() and \
-                not self.browser_tasks.has_tasks() and not \
-                self.needs_animation_frame and not \
-                self.pending_scroll and not \
-                self.needs_quit:
-                self.condition.wait()
-            self.lock.release()
-
-class TabWrapper:
-    def __init__(self, browser):
-        self.tab = Tab(self.commit)
-        self.browser = browser
-        self.url = None
-        self.scroll = 0
-
-    def schedule_load(self, url, body=None):
-        self.tab.main_thread_runner.schedule_browser_task(
-            Task(self.tab.load, url, body))
-        self.browser.set_needs_chrome_raster()
-
-    def commit(self, url, scroll, tab_bounds, display_list):
-        self.browser.compositor_lock.acquire(blocking=True)
-        if url != self.url or scroll != self.scroll:
-            self.browser.set_needs_chrome_raster()
-        self.url = url
-        if scroll != None:
-            self.scroll = scroll
-        self.browser.active_tab_bounds = tab_bounds
-        self.browser.active_tab_display_list = display_list.copy()
-        self.browser.set_needs_tab_raster()
-        self.browser.compositor_lock.release()
-
-    def schedule_click(self, x, y):
-        self.tab.main_thread_runner.schedule_browser_task(
-            Task(self.tab.click, x, y))
-
-    def schedule_keypress(self, char):
-        self.tab.main_thread_runner.schedule_browser_task(
-            Task(self.tab.keypress, char))
-
-    def schedule_go_back(self):
-        self.tab.main_thread_runner.schedule_browser_task(
-            Task(self.tab.go_back))
-
-    def schedule_scroll(self, scroll):
-        self.scroll = scroll
-        self.tab.main_thread_runner.schedule_scroll(scroll)
-
-    def handle_quit(self):
-        print("Time in style, layout and paint: {:>.6f}s".format(
-            self.tab.time_in_style_layout_and_paint))
-        self.tab.main_thread_runner.set_needs_quit()
-
-REFRESH_RATE_SEC = 0.016 # 16ms
-
-class PaintChunk:
-    def __init__(self, ancestor_effects):
-        self.ancestor_effects = ancestor_effects
-        self.chunk_items = []
-
-        self.composited_ancestor_index = -1
-        count = len(ancestor_effects) - 1
-        for display_item in reversed(ancestor_effects):
-            if display_item.needs_compositing():
-                self.composited_ancestor_index = count
-                break
-            count -= 1
-
-    def bounds(self):
-        retval = skia.Rect.MakeEmpty()
-        for item in self.chunk_items:
-            retval.join(item.bounds())
-        for display_item in reversed(self.ancestor_effects):
-            if display_item.needs_compositing():
-                break
-            retval = display_item.transform(retval)
-        return retval
-
-    def append(self, display_item):
-        self.chunk_items.append(display_item)
-
-    def needs_compositing(self):
-        return self.composited_ancestor_index >= 0
-
-    def composited_item(self):
-        if not self.needs_compositing():
-            return None
-        return self.ancestor_effects[self.composited_ancestor_index]
-
-    def display_list(self):
-        return self.chunk_items
-
-    def raster(self, canvas):
-        def op():
-            for display_item in self.chunk_items:
-                display_item.execute(canvas)
-        self.draw_internal(canvas, op, self.composited_ancestor_index + 1, False)
-
-    def draw_internal(self, canvas, op, index, is_raster):
-        if index == len(self.ancestor_effects):
-            op()
-        else:
-            display_item = self.ancestor_effects[index]
-            def recurse_op():
-                self.draw_internal(canvas, op, index + 1, is_raster)
-            display_item.draw(canvas, recurse_op, is_raster)
-
-    def draw(self, canvas, op, is_raster):
-        self.draw_internal(canvas, op, 0, is_raster)
-
-
-def display_list_to_paint_chunks_internal(
-    display_list, chunks, ancestor_effects):
-    current_chunk = None
-    for display_item in display_list:
-        if display_item.get_cmds() != None:
-            display_list_to_paint_chunks_internal(display_item.get_cmds(), chunks,
-                ancestor_effects + [display_item])
-        else:
-            if not current_chunk:
-                current_chunk = PaintChunk(ancestor_effects)
-                chunks.append(current_chunk)
-            current_chunk.append(display_item)
-
-def print_chunks(chunks):
-    for chunk in chunks:
-        print('chunk:')
-        print("  chunk display items:")
-        for display_item in chunk.chunk_items:
-            print(" " * 4 + str(display_item))
-        print("  chunk ancestor visual effect (skipping no-ops):")
-        count = 4
-        for display_item in chunk.ancestor_effects:
-            if not display_item.is_noop():
-                print(" " * count + str(display_item))
-                count += 2
-
-def display_list_to_paint_chunks(display_list):
-    chunks = []
-    display_list_to_paint_chunks_internal(display_list, chunks, [])
-    return chunks
-
-def print_composited_layers(composited_layers):
-    for layer in composited_layers:
-        print("layer: bounds=" + str(layer.bounds()))
-
-def do_composite(display_list, initial_layer):
-    chunks = display_list_to_paint_chunks(display_list)
-    composited_layers = [initial_layer]
-    for chunk in chunks:
-        placed = False
-        for layer in reversed(composited_layers):
-            if layer.can_merge(chunk):
-                layer.append(chunk)
-                placed = True
-                break
-            elif layer.overlaps(chunk.bounds()):
-                composited_layers.append(
-                    CompositedLayer(first_chunk=chunk))
-                placed = True
-                break
-        if not placed:            composited_layers.append(
-                CompositedLayer(first_chunk=chunk))
-    return composited_layers
-
-class Browser:
-    def __init__(self):
-        self.sdl_window = sdl2.SDL_CreateWindow(b"Browser",
-            sdl2.SDL_WINDOWPOS_CENTERED, sdl2.SDL_WINDOWPOS_CENTERED,
-            WIDTH, HEIGHT, sdl2.SDL_WINDOW_SHOWN)
-        self.root_surface = skia.Surface.MakeRaster(
-            skia.ImageInfo.Make(
-            WIDTH, HEIGHT,
-            ct=skia.kRGBA_8888_ColorType,
-            at=skia.kUnpremul_AlphaType))
-        self.chrome_surface = skia.Surface(WIDTH, CHROME_PX)
-        self.composited_layers = None
-
-        self.tabs = []
-        self.active_tab = None
-        self.focus = None
-        self.address_bar = ""
-        self.compositor_lock = threading.Lock()
-
-        self.time_in_raster_and_draw = 0
-        self.time_in_draw = 0
-
-        if sdl2.SDL_BYTEORDER == sdl2.SDL_BIG_ENDIAN:
-            self.RED_MASK = 0xff000000
-            self.GREEN_MASK = 0x00ff0000
-            self.BLUE_MASK = 0x0000ff00
-            self.ALPHA_MASK = 0x000000ff
-        else:
-            self.RED_MASK = 0x000000ff
-            self.GREEN_MASK = 0x0000ff00
-            self.BLUE_MASK = 0x00ff0000
-            self.ALPHA_MASK = 0xff000000
-
-        self.needs_tab_raster = False
-        self.needs_chrome_raster = True
-        self.needs_draw = True
-
-        self.active_tab_bound = None
-        self.active_tab_display_list = None
-        self.composited_layers = []
-        self.tab_surface = None
+        self.document.draw(self.display_list)
+        self.render()
+        self.max_y = self.document.h - HEIGHT
 
     def render(self):
-        assert not USE_BROWSER_THREAD
-        tab = self.tabs[self.active_tab].tab
-        tab.run_animation_frame()
-
-    def set_needs_tab_raster(self):
-        self.needs_tab_raster = True
-        self.needs_draw = True
-
-    def set_needs_chrome_raster(self):
-        self.needs_chrome_raster = True
-        self.needs_draw = True
-
-    def set_needs_draw(self):
-        self.needs_draw = True
-
-    def composite(self):
-        print('\n\ncompositing...')
-        initial_layer = CompositedLayer(bounds=self.active_tab_bounds)
-
-#        print_chunks(display_list_to_paint_chunks(self.active_tab_display_list))
-
-        self.composited_layers = do_composite(
-            self.active_tab_display_list, initial_layer)
-
-    def composite_raster_draw(self):
-        self.compositor_lock.acquire(blocking=True)
-        timer = None
-        draw_timer = None
-        if self.needs_draw:
-            timer = Timer()
-            timer.start()
-        if self.needs_chrome_raster:
-            self.raster_chrome()
-        if self.needs_tab_raster:
-            self.composite()
-            self.raster_tab()
-        if self.needs_draw:
-            draw_timer = Timer()
-            draw_timer.start()
-            self.draw()
-            self.time_in_draw += draw_timer.stop()
-        self.needs_tab_raster = False
-        self.needs_chrome_raster = False
-        self.needs_draw = False
-        self.compositor_lock.release()
-        if timer:
-            self.time_in_raster_and_draw += timer.stop()
-
-    def handle_down(self):
-        self.compositor_lock.acquire(blocking=True)
-        if not self.active_tab_bounds:
-            return
-        active_tab = self.tabs[self.active_tab]
-        active_tab.schedule_scroll(
-            clamp_scroll(
-                active_tab.scroll + SCROLL_STEP,
-                self.active_tab_bounds.height()))
-        self.set_needs_draw()
-        self.compositor_lock.release()
-
-    def handle_click(self, e):
-        self.compositor_lock.acquire(blocking=True)
-        if e.y < CHROME_PX:
-            self.focus = None
-            if 40 <= e.x < 40 + 80 * len(self.tabs) and 0 <= e.y < 40:
-                self.active_tab = int((e.x - 40) / 80)
-            elif 10 <= e.x < 30 and 10 <= e.y < 30:
-                self.load("https://browser.engineering/")
-            elif 10 <= e.x < 35 and 40 <= e.y < 90:
-                self.tabs[self.active_tab].schedule_go_back()
-            elif 50 <= e.x < WIDTH - 10 and 40 <= e.y < 90:
-                self.focus = "address bar"
-                self.address_bar = ""
-            self.set_needs_chrome_raster()
-        else:
-            self.focus = "content"
-            self.tabs[self.active_tab].schedule_click(
-                e.x, e.y - CHROME_PX)
-        self.compositor_lock.release()
-
-    def handle_key(self, char):
-        self.compositor_lock.acquire(blocking=True)
-        if not (0x20 <= ord(char) < 0x7f): return
+        self.timer.start("Rendering")
+        self.canvas.delete("all")
+        for cmd in self.display_list:
+            if cmd.y1 > self.scroll + HEIGHT - 60: continue
+            if cmd.y2 < self.scroll: continue
+            cmd.draw(self.scroll - 60, self.canvas)
+        self.timer.start("Chrome")
+        self.canvas.create_rectangle(0, 0, 800, 60, width=0, fill='light gray')
+        self.canvas.create_rectangle(50, 10, 790, 50)
+        font = tkinter.font.Font(family="Courier", size=30)
+        self.canvas.create_text(55, 15, anchor='nw', text=self.address_bar, font=font)
+        self.canvas.create_rectangle(10, 10, 35, 50)
+        self.canvas.create_polygon(15, 30, 30, 15, 30, 45, fill='black')
         if self.focus == "address bar":
-            self.address_bar += char
-            self.set_needs_chrome_raster()
-        elif self.focus == "content":
-            self.tabs[self.active_tab].schedule_keypress(char)
-        self.compositor_lock.release()
+            w = font.measure(self.address_bar)
+            self.canvas.create_line(55 + w, 15, 55 + w, 45)
+        elif isinstance(self.focus, InputLayout):
+            text = self.focus.node.attributes.get("value", "")
+            x = self.focus.x + self.focus.font.measure(text)
+            y = self.focus.y - self.scroll + 60
+            self.canvas.create_line(x, y, x, y + self.focus.h)
+        self.timer.stop()
 
-    def handle_enter(self):
-        self.compositor_lock.acquire(blocking=True)
-        if self.focus == "address bar`":
-            self.tabs[self.active_tab].schedule_load(self.address_bar)
-            self.tabs[self.active_tab].url = self.address_bar
-            self.focus = None
-            self.set_needs_chrome_raster()
-        self.compositor_lock.release()
-
-    def load(self, url):
-        new_tab = TabWrapper(self)
-        new_tab.schedule_load(url)
-        self.active_tab = len(self.tabs)
-        self.tabs.append(new_tab)
-
-    def raster_tab(self):
-        for composited_layer in self.composited_layers:
-            composited_layer.raster()
-
-    def raster_chrome(self):
-        canvas = self.chrome_surface.getCanvas()
-        canvas.clear(skia.ColorWHITE)
-    
-        # Draw the tabs UI:
-        tabfont = skia.Font(skia.Typeface('Arial'), 20)
-        for i, tab in enumerate(self.tabs):
-            name = "Tab {}".format(i)
-            x1, x2 = 40 + 80 * i, 120 + 80 * i
-            draw_line(canvas, x1, 0, x1, 40)
-            draw_line(canvas, x2, 0, x2, 40)
-            draw_text(canvas, x1 + 10, 10, name, tabfont)
-            if i == self.active_tab:
-                draw_line(canvas, 0, 40, x1, 40)
-                draw_line(canvas, x2, 40, WIDTH, 40)
-
-        # Draw the plus button to add a tab:
-        buttonfont = skia.Font(skia.Typeface('Arial'), 30)
-        draw_rect(canvas, 10, 10, 30, 30)
-        draw_text(canvas, 11, 4, "+", buttonfont)
-
-        # Draw the URL address bar:
-        draw_rect(canvas, 40, 50, WIDTH - 10, 90)
-        if self.focus == "address bar":
-            draw_text(canvas, 55, 55, self.address_bar, buttonfont)
-            w = buttonfont.measureText(self.address_bar)
-            draw_line(canvas, 55 + w, 55, 55 + w, 85)
-        else:
-            url = self.tabs[self.active_tab].url
-            if url:
-                draw_text(canvas, 55, 55, url, buttonfont)
-
-        # Draw the back button:
-        draw_rect(canvas, 10, 50, 35, 90)
-        path = \
-            skia.Path().moveTo(15, 70).lineTo(30, 55).lineTo(30, 85)
-        paint = skia.Paint(
-            Color=skia.ColorBLACK, Style=skia.Paint.kFill_Style)
-        canvas.drawPath(path, paint)
-
-    def draw(self):
-        canvas = self.root_surface.getCanvas()
-        canvas.clear(skia.ColorWHITE)
-        
-        draw_offset=(0, CHROME_PX - self.tabs[self.active_tab].scroll)
-        if self.composited_layers:
-            for composited_layer in self.composited_layers:
-                composited_layer.draw(canvas, draw_offset)
-
-        chrome_rect = skia.Rect.MakeLTRB(0, 0, WIDTH, CHROME_PX)
-        canvas.save()
-#        canvas.clipRect(chrome_rect)
-        self.chrome_surface.draw(canvas, 0, 0)
-        canvas.restore()
-
-        # This makes an image interface to the Skia surface, but
-        # doesn't actually copy anything yet.
-        skia_image = self.root_surface.makeImageSnapshot()
-        skia_bytes = skia_image.tobytes()
-
-        depth = 32 # Bits per pixel
-        pitch = 4 * WIDTH # Bytes per row
-        sdl_surface = sdl2.SDL_CreateRGBSurfaceFrom(
-            skia_bytes, WIDTH, HEIGHT, depth, pitch,
-            self.RED_MASK, self.GREEN_MASK,
-            self.BLUE_MASK, self.ALPHA_MASK)
-
-        rect = sdl2.SDL_Rect(0, 0, WIDTH, HEIGHT)
-        window_surface = sdl2.SDL_GetWindowSurface(self.sdl_window)
-        # SDL_BlitSurface is what actually does the copy.
-        sdl2.SDL_BlitSurface(sdl_surface, rect, window_surface, rect)
-        sdl2.SDL_UpdateWindowSurface(self.sdl_window)
-
-    def handle_quit(self):
-        print("Time in raster and draw: {:>.6f}s".format(
-            self.time_in_raster_and_draw))
-        print("Time in draw: {:>.6f}s".format(
-            self.time_in_draw))
-
-        self.tabs[self.active_tab].handle_quit()
-        sdl2.SDL_DestroyWindow(self.sdl_window)
+    def scrolldown(self, e):
+        self.scroll = self.scroll + SCROLL_STEP
+        self.scroll = min(self.scroll, self.max_y)
+        self.scroll = max(0, self.scroll)
+        self.render()
 
 if __name__ == "__main__":
     import sys
-
-    sdl2.SDL_Init(sdl2.SDL_INIT_EVENTS)
     browser = Browser()
     browser.load(sys.argv[1])
-
-    event = sdl2.SDL_Event()
-    while True:
-        if sdl2.SDL_PollEvent(ctypes.byref(event)) != 0:
-            if event.type == sdl2.SDL_QUIT:
-                browser.handle_quit()
-                sdl2.SDL_Quit()
-                sys.exit()
-                break
-            elif event.type == sdl2.SDL_MOUSEBUTTONUP:
-                browser.handle_click(event.button)
-            elif event.type == sdl2.SDL_KEYDOWN:
-                if event.key.keysym.sym == sdl2.SDLK_RETURN:
-                    browser.handle_enter()
-                elif event.key.keysym.sym == sdl2.SDLK_DOWN:
-                    browser.handle_down()
-            elif event.type == sdl2.SDL_TEXTINPUT:
-                browser.handle_key(event.text.text.decode('utf8'))
-        if not USE_BROWSER_THREAD and \
-            browser.tabs[browser.active_tab].tab.display_scheduled:
-            browser.render()
-        browser.composite_raster_draw()
+    tkinter.mainloop()
